refactor: リクエストボディの型をroute.tsへ移動
