+++ src/app/_types/pet/CreatePet.ts
@@ -1,13 +1,3 @@
-//クライアント→APIに送られてくるデータ（リクエスト）
-export type CreatePetRequest = {
-  name: string;
-  species: string;
-  gender: "おとこのこ♂" | "おんなのこ♀" | null;
-  birthday: string | null;
-  adoptedAt: string | null;
-  imagePath?: string;
-};
-
 //サーバーから返ってくるデータ（レスポンス）
 export type Pet = {
   id: number;
